fix: Unified utils files
