+++ packages/kilo-vscode/webview-ui/src/components/chat/permission-dock-utils.ts
@@ -17,3 +17,54 @@ export function savedRuleStates(rules: string[], saved: PermissionRule | undefin
   }
   return result
 }
+
+// ---------------------------------------------------------------------------
+// Human-readable permission descriptions
+// ---------------------------------------------------------------------------
+
+/** Maps tool names to their i18n key for the human-readable label. */
+export const TOOL_LABEL_KEYS: Record<string, string> = {
+  read: "ui.permission.toolLabel.read",
+  edit: "ui.permission.toolLabel.edit",
+  write: "ui.permission.toolLabel.write",
+  patch: "ui.permission.toolLabel.patch",
+  multiedit: "ui.permission.toolLabel.edit",
+  glob: "ui.permission.toolLabel.globSearch",
+  grep: "ui.permission.toolLabel.grepSearch",
+  list: "ui.permission.toolLabel.list",
+  external_directory: "ui.permission.toolLabel.externalDirectory",
+  webfetch: "ui.permission.toolLabel.webFetch",
+  websearch: "ui.permission.toolLabel.webSearch",
+  task: "ui.permission.toolLabel.task",
+  skill: "ui.permission.toolLabel.skill",
+  lsp: "ui.permission.toolLabel.lsp",
+}
+
+export type PatternDescription = { kind: "single"; text: string } | { kind: "multi"; title: string; paths: string[] }
+
+/** Resolve the human-readable label for a tool (e.g. "Read", "Web Fetch"). */
+export function resolveLabel(tool: string, t: (key: string) => string): string {
+  const key = TOOL_LABEL_KEYS[tool]
+  return key ? t(key) : tool
+}
+
+/**
+ * Build a human-readable description for a permission request's patterns.
+ *
+ * Returns null when there are no meaningful patterns to display (e.g. only "*").
+ * For a single pattern: "Read src/app.ts"
+ * For multiple patterns: { title: "Read:", paths: ["src/app.ts", "src/index.ts"] }
+ */
+export function describePatterns(
+  tool: string,
+  patterns: string[],
+  t: (key: string) => string,
+): PatternDescription | null {
+  const filtered = patterns.filter((p) => p !== "*")
+  if (filtered.length === 0) return null
+
+  const key = TOOL_LABEL_KEYS[tool]
+  const label = key ? t(key) : tool
+  if (filtered.length === 1) return { kind: "single", text: `${label} ${filtered[0]}` }
+  return { kind: "multi", title: `${label}:`, paths: filtered }
+}
